requirements.txt
python-dotenv==0.10.1
flask==1.0.3
Flask-SQLAlchemy==2.4
Flask-Admin==1.5.3
Flask-BasicAuth==0.2.0
speaklater==1.3
pytelegrambotapi==2.3.0
